hw4 SQlite/requirements/dev.txt
certifi==2022.12.7
charset-normalizer==3.0.1
Currency==0.1.2
CurrencyConverter==0.17.5
idna==3.4
ijson==3.2.0.post0
oauthlib==3.2.2
python-iso4217==0.0.8
requests==2.28.2
requests-oauthlib==1.3.1
urllib3==1.26.14